Implement "verbose" CLI option for logging
If the `-v` option is specified on the CLI, NoiseTorch will print logs
to stderr. If not, log output will be silently discarded.

Handling of this is done in `cli.go` instead of in `main.go` where the
TODO was listed for it. This makes more sense, as handling of CLI
options should be done as much as possible within the `doCLI` function,
not in `main`.