Add short title to capabilities view
Adds a shorter main title to the capabilities view which can be centered
and not wrap. The previous title is displayed below, but is left-aligned
and can wrap.

Also changes "I'll fix it for you" to "We'll fix it..." because the
previous sentence says "We require..." and not "I require..."